first versione, not all the functionality

+++ src/flagsControll.c
@@ -7,7 +7,7 @@ int controllFlags(int argc, char **args){
     int position = -1;
     for(int i = 1; i < argc; i++){
         if(args[i][0] == '-'){
-            for(int j = 1; j < strlen(args[i]); j++){
+            for(size_t j = 1; j < strlen(args[i]); j++){
                flagsOn(args[i][j]); 
             }
         }else{
